moved environment.ts to utils file
